HT_21/requirements.txt
Django==3.2.12
djangorestframework==3.12.4
django-cors-headers==3.10.0